+++ pkg/moonshine/lib.go
@@ -147,6 +147,8 @@ var (
 	fnStopStream                 func(transcriberHandle, streamHandle int32) int32
 	fnAddAudioToStream           func(transcriberHandle, streamHandle int32, newAudioData *float32, audioLength uint64, sampleRate int32, flags uint32) int32
 	fnTranscribeStream           func(transcriberHandle, streamHandle int32, flags uint32, outTranscript *unsafe.Pointer) int32
+	fnTranscriberSetKeyterms     func(transcriberHandle int32, keyterms *byte) int32
+	fnTranscriberSetContext      func(transcriberHandle int32, context *byte, maxTerms int32) int32
 
 	fnCreateTTSSynthesizerFromFiles  func(language string, filenames unsafe.Pointer, filenamesCount uint64, options *cOption, optionsCount uint64, version int32) int32
 	fnCreateTTSSynthesizerFromMemory func(language string, filenames unsafe.Pointer, filenamesCount uint64, memory unsafe.Pointer, memorySizes unsafe.Pointer, options *cOption, optionsCount uint64, version int32) int32
@@ -181,6 +183,8 @@ func registerSymbols(h uintptr) {
 	reg(&fnStopStream, "moonshine_stop_stream")
 	reg(&fnAddAudioToStream, "moonshine_transcribe_add_audio_to_stream")
 	reg(&fnTranscribeStream, "moonshine_transcribe_stream")
+	reg(&fnTranscriberSetKeyterms, "moonshine_transcriber_set_keyterms")
+	reg(&fnTranscriberSetContext, "moonshine_transcriber_set_context")
 
 	reg(&fnCreateTTSSynthesizerFromFiles, "moonshine_create_tts_synthesizer_from_files")
 	reg(&fnCreateTTSSynthesizerFromMemory, "moonshine_create_tts_synthesizer_from_memory")